Add TuoteService, update TuoteController and TuoteRepository, get kahvilaitteet.html page show some precreated product examples

+++ src/main/java/projekti/kahvikauppa/Tuote.java
@@ -2,9 +2,7 @@ package projekti.kahvikauppa;
 
 import java.math.BigDecimal;
 import javax.persistence.Entity;
-
 import org.springframework.data.jpa.domain.AbstractPersistable;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class Tuote extends AbstractPersistable<Long> {
     private String nimi;
     private String kuvaus;
-    private BigDecimal hinta;
+    private Integer hinta;
     private String kuva;
 }
